# openapi.yaml
openapi: "3.1.0"
info:
  title: Quest Board API
  version: "0.1.0"
  description: |
    Backend API for Quest Board. TypeScript types for all schemas
    are auto-generated from Rust structs via ts-rs in
    `frontend/src/lib/types/bindings/`.
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User registered successfully
        "400":
          description: Invalid input or username already registered
  /auth/login:
    post:
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
        "401":
          description: Not logged in

  # ── Boards ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
        name: { type: string }
    LoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
...